modify output filenames to apply OOP for catalysis
